Time: 35 ms (78.34%), Space: 104.5 MB (16.48%) - LeetHub

+++ 2300-successful-pairs-of-spells-and-potions/2300-successful-pairs-of-spells-and-potions.cpp
@@ -7,7 +7,7 @@ public:
             int left = 0, right = potions.size() - 1;
             while (left <= right) {
                 int mid = left + (right - left) / 2;
-                long long product = spells[i] * potions[mid];
+                long long product = (long long)spells[i] * potions[mid];
                 if (product < success) {
                     left = mid + 1;
                 } else {
